+++ src/modules/album/album.module.ts
@@ -6,16 +6,14 @@ import { ArtistService } from '../artist/artist.service';
 import { BandService } from '../band/band.service';
 import { GenreService } from '../genre/genre.service';
 import { TrackService } from '../track/track.service';
+import { ReferenceService } from 'src/reference/reference.service';
 
 @Module({
   imports: [HttpModule],
   providers: [
     AlbumService,
     AlbumResolver,
-    ArtistService,
-    BandService,
-    GenreService,
-    TrackService,
+    ReferenceService,
   ],
   exports: [AlbumService],
 })
