+++ src/modules/band/band.module.ts
@@ -3,9 +3,10 @@ import { BandService } from './band.service';
 import { BandResolver } from './band.resolver';
 import { HttpModule } from '@nestjs/axios';
 import { GenreService } from '../genre/genre.service';
+import { ReferenceService } from 'src/reference/reference.service';
 
 @Module({
   imports: [HttpModule],
-  providers: [BandService, BandResolver, GenreService],
+  providers: [BandService, BandResolver,ReferenceService],
 })
 export class BandModule {}
