+++ src/urls/urls.ts
@@ -2,7 +2,7 @@ export enum Path {
   album = 'http://localhost:3005/v1/albums/',
   artist = 'http://localhost:3002/v1/artists/',
   band = 'http://localhost:3003/v1/bands/',
-  fav = 'http://localhost:3007/v1/favourites',
+  fav = 'http://localhost:3007/v1/favourites/',
   genre = 'http://localhost:3001/v1/genres/',
   track = 'http://localhost:3006/v1/tracks/',
 }
